Move test code of example app to the basic example.
